# server/api/admin/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Administration API for managing Magec multi-agent voice assistant resources.\nWhen server.adminPassword is configured, all /api/ endpoints require a Bearer token.",
        "title": "Magec Admin API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1/admin",
    "paths": {
...
        "/backends": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Returns all configured LLM/TTS/transcription backends",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backends"
                ],
                "summary": "List backends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.BackendDefinition"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Creates a new LLM/TTS/transcription backend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backends"
                ],
                "summary": "Create backend",
                "parameters": [
                    {
                        "description": "Backend definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.BackendDefinition"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.BackendDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/admin.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/backends/{id}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Returns a backend by its unique ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backends"
                ],
                "summary": "Get backend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backend ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.BackendDefinition"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Updates a backend by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backends"
                ],
                "summary": "Update backend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backend ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Backend definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.BackendDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.BackendDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Deletes a backend by ID",
                "tags": [
                    "backends"
                ],
                "summary": "Delete backend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backend ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource not found"
                }
            }
        },
...
        "store.BackendDefinition": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Enter \"Bearer {adminPassword}\" to authenticate. Only required when server.adminPassword is set.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}